Upgrade project to work with SilverStripe 4
Adds much needed tests to ensure functionality remains as intended,
and takes care of some code cleanliness. There is now heavy documentation
to help serve the secondary purpose of a learning tool for interested parties.

One caveat is that the tests rely on a feature not yet merged into core, that
of being able to apply extensions via Injector services. The extension will
remain functional however, only some of the test suite may fail.

A new feature has also been introduced in the way of an extension for the
controller portion of the slug handling - instead of relying on commented
examples in the config file.

This new feature also allows for having multiple associations on a Controller
to various slugged objects, and is flexible in whether or not a slugged item
is available as the action itself, or under a different name denoting different
relationships to the data.